+++ tests/export_test.py
@@ -1,4 +1,3 @@
-import re
 from xml.etree import ElementTree as ET
 
 import pytest
@@ -28,18 +27,20 @@ def test_export_1_1(datadir, tmp_path):
     tmpdir = tmp_path / "test_export_1_1"
     tmpdir.mkdir()
     tmppath = tmpdir / "mini_lmf_export_1_1.xml"
-    lexicons = wn.lexicons(lexicon="test-ja")
+    lexicons = wn.lexicons(lexicon="test-ja test-en-ext")
     wn.export(lexicons, tmppath, version="1.1")
 
     # remove comments, indentation, etc.
     orig = ET.canonicalize(from_file=datadir / "mini-lmf-1.1.xml", strip_text=True)
-    # temporary until support for exporting extensions
-    orig = re.sub(r"\s*<LexiconExtension.*</LexiconExtension>\s*", "", orig, flags=re.M)
     temp = ET.canonicalize(from_file=tmppath, strip_text=True)
     # additional transformation to help with debugging
     orig = orig.replace("<", "\n<")
     temp = temp.replace("<", "\n<")
     assert orig == temp
+
+    # fails when exporting to WN-LMF 1.0
+    with pytest.raises(wn.Error):
+        wn.export(lexicons, tmppath, version="1.0")
 
 
 @pytest.mark.usefixtures("mini_db_1_4")
@@ -47,13 +48,11 @@ def test_export_1_4(datadir, tmp_path):
     tmpdir = tmp_path / "test_export_1_4"
     tmpdir.mkdir()
     tmppath = tmpdir / "mini_lmf_export_1_4.xml"
-    lexicons = wn.lexicons(lexicon="test-1.4")
+    lexicons = wn.lexicons(lexicon="test-1.4 test-ext-1.4")
     wn.export(lexicons, tmppath, version="1.4")
 
     # remove comments, indentation, etc.
     orig = ET.canonicalize(from_file=datadir / "mini-lmf-1.4.xml", strip_text=True)
-    # temporary until support for exporting extensions
-    orig = re.sub(r"\s*<LexiconExtension.*</LexiconExtension>\s*", "", orig, flags=re.M)
     temp = ET.canonicalize(from_file=tmppath, strip_text=True)
     # additional transformation to help with debugging
     orig = orig.replace("<", "\n<")