+++ wn/_queries.py
@@ -913,6 +913,38 @@ def get_sense_synset_relations(
     yield from rows
 
 
+def get_relation_targets(
+    rel_table: str,
+    tgt_table: str,
+    lexicons: Sequence[str],
+    target_lexicons: Sequence[str],
+) -> set[str]:
+    if rel_table not in {
+        "sense_relations",
+        "sense_synset_relations",
+        "synset_relations",
+    }:
+        raise ValueError(f"invalid relation table: {rel_table}")
+    if tgt_table not in ("senses", "synsets"):
+        raise ValueError(f"invalid target table: {tgt_table}")
+    params: list = [*lexicons, *target_lexicons]
+    query = f"""
+        WITH
+          lexrowids(rowid) AS
+            (SELECT rowid FROM lexicons WHERE specifier IN ({_vs(lexicons)})),
+          tgtlexrowids(rowid) AS
+            (SELECT rowid FROM lexicons WHERE specifier IN ({_vs(target_lexicons)}))
+        SELECT DISTINCT tgt.id
+          FROM {rel_table} AS srel
+          JOIN lexicons AS lex ON srel.lexicon_rowid = lex.rowid
+          JOIN {tgt_table} AS tgt ON tgt.rowid = srel.target_rowid
+         WHERE srel.lexicon_rowid IN lexrowids
+           AND tgt.lexicon_rowid IN tgtlexrowids
+    """
+    rows: Iterator[str] = connect().execute(query, params)
+    return {row[0] for row in rows}
+
+
 _SANITIZED_METADATA_TABLES = {
     # 'ilis': 'ilis',
     # 'proposed_ilis': 'proposed_ilis',